+++ src/main/java/shadows/apotheosis/deadly/DeadlyModuleEvents.java
@@ -4,30 +4,22 @@ import java.util.Map;
 import java.util.Set;
 
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Multimap;
-import com.google.common.collect.TreeMultimap;
 
-import net.minecraft.world.entity.ai.attributes.Attribute;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.event.ItemAttributeModifierEvent;
 import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.event.entity.living.LivingEntityUseItemEvent;
-import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.loading.FMLEnvironment;
-import net.minecraftforge.fml.util.ObfuscationReflectionHelper;
 import shadows.apotheosis.Apoth;
+import shadows.apotheosis.deadly.affix.Affix;
+import shadows.apotheosis.deadly.affix.AffixHelper;
 import shadows.apotheosis.deadly.commands.CategoryCheckCommand;
 import shadows.apotheosis.deadly.commands.LootifyCommand;
 import shadows.apotheosis.deadly.commands.ModifierCommand;
 import shadows.apotheosis.deadly.commands.RarityCommand;
 import shadows.apotheosis.deadly.loot.LootCategory;
-import shadows.apotheosis.deadly.loot.affix.Affix;
-import shadows.apotheosis.deadly.loot.affix.AffixHelper;
 
 public class DeadlyModuleEvents {
 
@@ -41,21 +33,6 @@ public class DeadlyModuleEvents {
 		CategoryCheckCommand.register(e.getDispatcher());
 		LootifyCommand.register(e.getDispatcher());
 		ModifierCommand.register(e.getDispatcher());
-	}
-
-	@SubscribeEvent(priority = EventPriority.LOWEST)
-	public void sortModifiers(ItemAttributeModifierEvent e) {
-		if (e.getModifiers() == null || e.getModifiers().isEmpty() || FMLEnvironment.dist == Dist.DEDICATED_SERVER) return;
-		Multimap<Attribute, AttributeModifier> map = TreeMultimap.create((k1, k2) -> k1.getRegistryName().compareTo(k2.getRegistryName()), (v1, v2) -> {
-			int compOp = Integer.compare(v1.getOperation().ordinal(), v2.getOperation().ordinal());
-			int compValue = Double.compare(v2.getAmount(), v1.getAmount());
-			return compOp == 0 ? compValue == 0 ? v1.getName().compareTo(v2.getName()) : compValue : compOp;
-		});
-		for (Map.Entry<Attribute, AttributeModifier> ent : e.getModifiers().entries()) {
-			if (ent.getKey() != null && ent.getValue() != null) map.put(ent.getKey(), ent.getValue());
-			else DeadlyModule.LOGGER.debug("Detected broken attribute modifier entry on item {}.  Attr={}, Modif={}", e.getItemStack(), ent.getKey(), ent.getValue());
-		}
-		ObfuscationReflectionHelper.setPrivateValue(ItemAttributeModifierEvent.class, e, map, "unmodifiableModifiers");
 	}
 
 	@SubscribeEvent