+++ src/main/java/shadows/apotheosis/ench/table/EnchantingRecipe.java
@@ -20,7 +20,7 @@ import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.common.crafting.CraftingHelper;
 import net.minecraftforge.registries.ForgeRegistryEntry;
-import shadows.apotheosis.ench.EnchModule;
+import shadows.apotheosis.Apoth.RecipeTypes;
 import shadows.apotheosis.ench.table.EnchantingStatManager.Stats;
 
 public class EnchantingRecipe implements Recipe<Container> {
@@ -105,7 +105,7 @@ public class EnchantingRecipe implements Recipe<Container> {
 
 	@Override
 	public RecipeType<?> getType() {
-		return EnchModule.INFUSION;
+		return RecipeTypes.INFUSION;
 	}
 
 	public static class Serializer extends ForgeRegistryEntry<RecipeSerializer<?>> implements RecipeSerializer<EnchantingRecipe> {
@@ -148,7 +148,7 @@ public class EnchantingRecipe implements Recipe<Container> {
 
 	@Nullable
 	public static EnchantingRecipe findMatch(Level level, ItemStack input, float eterna, float quanta, float arcana) {
-		List<EnchantingRecipe> recipes = level.getRecipeManager().getAllRecipesFor(EnchModule.INFUSION);
+		List<EnchantingRecipe> recipes = level.getRecipeManager().getAllRecipesFor(RecipeTypes.INFUSION);
 		recipes.sort((r1, r2) -> -Float.compare(r1.requirements.eterna, r2.requirements.eterna));
 		for (EnchantingRecipe r : recipes)
 			if (r.matches(input, eterna, quanta, arcana)) return r;
@@ -156,7 +156,7 @@ public class EnchantingRecipe implements Recipe<Container> {
 	}
 
 	public static EnchantingRecipe findItemMatch(Level level, ItemStack toEnchant) {
-		return level.getRecipeManager().getAllRecipesFor(EnchModule.INFUSION).stream().filter(r -> r.getInput().test(toEnchant)).findFirst().orElse(null);
+		return level.getRecipeManager().getAllRecipesFor(RecipeTypes.INFUSION).stream().filter(r -> r.getInput().test(toEnchant)).findFirst().orElse(null);
 	}
 
 }