+++ src/main/java/shadows/apotheosis/spawn/modifiers/SpawnerModifier.java
@@ -20,7 +20,7 @@ import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.registries.ForgeRegistryEntry;
-import shadows.apotheosis.spawn.SpawnerModule;
+import shadows.apotheosis.Apoth.RecipeTypes;
 import shadows.apotheosis.spawn.spawner.ApothSpawnerTile;
 
 /**
@@ -124,12 +124,12 @@ public class SpawnerModifier implements Recipe<Container> {
 
 	@Override
 	public RecipeType<?> getType() {
-		return SpawnerModule.MODIFIER;
+		return RecipeTypes.MODIFIER;
 	}
 
 	@Nullable
 	public static SpawnerModifier findMatch(ApothSpawnerTile tile, ItemStack mainhand, ItemStack offhand) {
-		List<SpawnerModifier> recipes = tile.getLevel().getRecipeManager().getAllRecipesFor(SpawnerModule.MODIFIER);
+		List<SpawnerModifier> recipes = tile.getLevel().getRecipeManager().getAllRecipesFor(RecipeTypes.MODIFIER);
 		recipes.sort((r1, r2) -> r1.offHand == Ingredient.EMPTY ? r2.offHand == Ingredient.EMPTY ? 0 : 1 : -1);
 		for (SpawnerModifier r : recipes)
 			if (r.matches(tile, mainhand, offhand)) return r;